port forward r572031 from 1.1 branch. fixes for DIRMINA-429 and DIRMINA-430
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@572512 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mina/common/DummySession.java
@@ -6,16 +6,16 @@
  *  to you under the Apache License, Version 2.0 (the
  *  "License"); you may not use this file except in compliance
  *  with the License.  You may obtain a copy of the License at
- *  
+ *
  *    http://www.apache.org/licenses/LICENSE-2.0
- *  
+ *
  *  Unless required by applicable law or agreed to in writing,
  *  software distributed under the License is distributed on an
  *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
  *  KIND, either express or implied.  See the License for the
  *  specific language governing permissions and limitations
- *  under the License. 
- *  
+ *  under the License.
+ *
  */
 package org.apache.mina.common;
 
@@ -25,17 +25,17 @@ import java.net.SocketAddress;
 /**
  * A dummy {@link IoSession} for unit-testing or non-network-use of
  * the classes that depends on {@link IoSession}.
- * 
+ *
  * @author The Apache MINA Project (dev@mina.apache.org)
  * @version $Rev$, $Date$
  */
 public class DummySession extends AbstractIoSession {
-    
+
     private static final TransportMetadata TRANSPORT_METADATA =
-        new DefaultTransportMetadata(
-                "dummy", false, false,
-                SocketAddress.class, IoSessionConfig.class, Object.class);
-    
+            new DefaultTransportMetadata(
+                    "dummy", false, false,
+                    SocketAddress.class, IoSessionConfig.class, Object.class);
+
     private static final SocketAddress ANONYMOUS_ADDRESS = new SocketAddress() {
         private static final long serialVersionUID = -496112902353454179L;
 
@@ -44,7 +44,7 @@ public class DummySession extends AbstractIoSession {
             return "?";
         }
     };
-    
+
     private volatile IoService service;
 
     private volatile IoSessionConfig config = new AbstractIoSessionConfig() {
@@ -52,7 +52,7 @@ public class DummySession extends AbstractIoSession {
         protected void doSetAll(IoSessionConfig config) {
         }
     };
-    
+
     private volatile IoFilterChain filterChain = new AbstractIoFilterChain(this) {
         @Override
         protected void doClose(IoSession session) throws Exception {
@@ -64,7 +64,7 @@ public class DummySession extends AbstractIoSession {
             fireMessageSent(session, writeRequest);
         }
     };
-    
+
     private volatile IoHandler handler = new IoHandlerAdapter();
     private volatile SocketAddress localAddress = ANONYMOUS_ADDRESS;
     private volatile SocketAddress remoteAddress = ANONYMOUS_ADDRESS;
@@ -100,14 +100,14 @@ public class DummySession extends AbstractIoSession {
                 return TRANSPORT_METADATA;
             }
         };
-        
+
         // Set meaningless default values.
         acceptor.setHandler(new IoHandlerAdapter());
         acceptor.setLocalAddress(ANONYMOUS_ADDRESS);
 
         this.service = acceptor;
     }
-    
+
     @Override
     protected void updateTrafficMask() {
     }
@@ -115,7 +115,7 @@ public class DummySession extends AbstractIoSession {
     public IoSessionConfig getConfig() {
         return config;
     }
-    
+
     /**
      * Sets the configuration of this session.
      */
@@ -123,14 +123,14 @@ public class DummySession extends AbstractIoSession {
         if (config == null) {
             throw new NullPointerException("config");
         }
-        
+
         this.config = config;
     }
 
     public IoFilterChain getFilterChain() {
         return filterChain;
     }
-    
+
     /**
      * Sets the filter chain that affects this session.
      */
@@ -138,14 +138,14 @@ public class DummySession extends AbstractIoSession {
         if (filterChain == null) {
             throw new NullPointerException("filterChain");
         }
-        
+
         this.filterChain = filterChain;
     }
 
     public IoHandler getHandler() {
         return handler;
     }
-    
+
     /**
      * Sets the {@link IoHandler} which handles this session.
      */
@@ -153,7 +153,7 @@ public class DummySession extends AbstractIoSession {
         if (handler == null) {
             throw new NullPointerException("handler");
         }
-        
+
         this.handler = handler;
     }
 
@@ -164,7 +164,7 @@ public class DummySession extends AbstractIoSession {
     public SocketAddress getRemoteAddress() {
         return remoteAddress;
     }
-    
+
     /**
      * Sets the socket address of local machine which is associated with
      * this session.
@@ -173,22 +173,22 @@ public class DummySession extends AbstractIoSession {
         if (localAddress == null) {
             throw new NullPointerException("localAddress");
         }
-        
+
         this.localAddress = localAddress;
     }
-    
+
     /**
-     * Sets the socket address of remote peer. 
+     * Sets the socket address of remote peer.
      */
     public void setRemoteAddress(SocketAddress remoteAddress) {
         if (remoteAddress == null) {
             throw new NullPointerException("remoteAddress");
         }
-        
+
         this.remoteAddress = remoteAddress;
     }
 
-    public int getScheduledWriteBytes() {
+    public long getScheduledWriteBytes() {
         return 0;
     }
 
@@ -199,7 +199,7 @@ public class DummySession extends AbstractIoSession {
     public IoService getService() {
         return service;
     }
-    
+
     /**
      * Sets the {@link IoService} which provides I/O service to this session.
      */
@@ -207,14 +207,14 @@ public class DummySession extends AbstractIoSession {
         if (service == null) {
             throw new NullPointerException("service");
         }
-        
+
         this.service = service;
     }
 
     public TransportMetadata getTransportMetadata() {
         return transportMetadata;
     }
-    
+
     /**
      * Sets the {@link TransportMetadata} that this session runs on.
      */
@@ -222,7 +222,7 @@ public class DummySession extends AbstractIoSession {
         if (transportMetadata == null) {
             throw new NullPointerException("transportMetadata");
         }
-        
+
         this.transportMetadata = transportMetadata;
     }
 }
